+++ lib/analyticsScoutCompletion/analyticsScoutCompletionView.dart
@@ -4,7 +4,7 @@ import 'package:cubook/model/class.dart';
 import 'package:cubook/model/task.dart';
 import 'package:cubook/model/themeInfo.dart';
 import 'package:cubook/analyticsScoutCompletion/analyticsScoutCompletionModel.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_view.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_view.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
