+++ lib/checkScoutTaskDetail/taskDetailScoutConfirm_view.dart
@@ -2,7 +2,7 @@ import 'package:chewie/chewie.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cubook/model/task.dart';
 import 'package:cubook/model/themeInfo.dart';
-import 'package:cubook/task_detail_scout_confirm/widget/taskDetailScoutConfirm_add.dart';
+import 'package:cubook/checkScoutTaskDetail/widget/taskDetailScoutConfirm_add.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
