+++ lib/checkScoutTaskDetail/widget/taskDetailScoutConfirm_add.dart
@@ -1,6 +1,6 @@
 import 'package:cubook/model/task.dart';
 import 'package:cubook/model/themeInfo.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_model.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_model.dart';
 import 'package:expand_widget/expand_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
