+++ lib/detailTaskWaiting/detrailTaskWaiting_view.dart
@@ -4,7 +4,7 @@ import 'package:cubook/detailTaskWaiting/detailTaskWaiting_model.dart';
 import 'package:cubook/model/class.dart';
 import 'package:cubook/model/task.dart';
 import 'package:cubook/model/themeInfo.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_view.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
@@ -39,8 +39,9 @@ class DetailTaskWaitingView_old extends StatelessWidget {
                   onTap: () {
                     FocusScope.of(context).unfocus();
                   },
-                  child: Scrollbar(child:SingleChildScrollView(
-                      child: Column(
+                  child: Scrollbar(
+                      child: SingleChildScrollView(
+                          child: Column(
                     children: <Widget>[
                       Hero(
                           tag: 'detailTask' + documentID!,
