+++ lib/home/homeModel.dart
@@ -1,11 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:cubook/homeLeader/homeLeader_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 // import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:google_sign_in/google_sign_in.dart';
 
+import 'homeLeader/leaderHomeView.dart';
 import 'homeScout/scoutHomeView.dart';
 
 class HomeModel extends ChangeNotifier {
