+++ lib/main.dart
@@ -1,7 +1,12 @@
-import 'package:cubook/createActivity/addLumpSelectItem/addLumpSelectItem_model.dart';
-import 'package:cubook/createActivity/addLumpSelectItem/addLumpSelectItem_view.dart';
 import 'package:cubook/analytics/analyticsModel.dart';
 import 'package:cubook/analytics/analyticsView.dart';
+import 'package:cubook/analyticsScoutCompletion/analyticsScoutCompletionModel.dart';
+import 'package:cubook/analyticsScoutCompletion/analyticsScoutCompletionView.dart';
+import 'package:cubook/analyticsTaskDetail/taskDetailAnalytics_model.dart';
+import 'package:cubook/analyticsTaskList/analyticsTaskListModel.dart';
+import 'package:cubook/checkScoutTaskList/taskListScoutConfirm_model.dart';
+import 'package:cubook/createActivity/addLumpSelectItem/addLumpSelectItem_model.dart';
+import 'package:cubook/createActivity/addLumpSelectItem/addLumpSelectItem_view.dart';
 import 'package:cubook/createActivity/createActivity_model.dart';
 import 'package:cubook/createActivity/createActivity_view.dart';
 import 'package:cubook/detailActivity/detailActivity_model.dart';
@@ -31,6 +36,7 @@ import 'package:cubook/listMigrationWaiting/listMigrationWaiting_view.dart';
 import 'package:cubook/listTaskWaiting/listTaskWaiting_model.dart';
 import 'package:cubook/listTaskWaiting/listTaskWaiting_view.dart';
 import 'package:cubook/notification/notification_model.dart';
+import 'package:cubook/scoutTaskList/scoutTaskListModel.dart';
 import 'package:cubook/settingAccount/settingAccount_model.dart';
 import 'package:cubook/settingAccount/settingAccount_view.dart';
 import 'package:cubook/settingAccount/widget/changeMailAddress.dart';
@@ -44,17 +50,10 @@ import 'package:cubook/settingGroupAccount/widget/changeName.dart';
 import 'package:cubook/settingGroupAccount/widget/deleteGroupAccount.dart';
 import 'package:cubook/settingGroupAccount/widget/editProfile.dart';
 import 'package:cubook/signup/signup_model.dart';
-import 'package:cubook/analyticsTaskDetail/taskDetailAnalytics_model.dart';
-import 'package:cubook/analyticsScoutCompletion/analyticsScoutCompletionModel.dart';
-import 'package:cubook/analyticsScoutCompletion/analyticsScoutCompletionView.dart';
-import 'package:cubook/analyticsTaskList/analyticsTaskListModel.dart';
-import 'package:cubook/scoutTaskList/scoutTaskListModel.dart';
-import 'package:cubook/task_list_scout_confirm/taskListScoutConfirm_model.dart';
 import 'package:cubook/userDetail/userDetail_model.dart';
 import 'package:dynamic_color/dynamic_color.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
-// import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
