+++ lib/model/class.dart
@@ -4,8 +4,8 @@ import 'package:cubook/model/themeInfo.dart';
 import 'package:cubook/scoutTaskDetail/scoutTaskModel.dart';
 import 'package:cubook/scoutTaskDetail/scoutTaskOverview.dart';
 import 'package:cubook/scoutTaskDetail/scoutTaskRecord.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_model.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_view.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_model.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
