+++ lib/userDetail/widget/selectBook.dart
@@ -2,21 +2,19 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cubook/model/class.dart';
 import 'package:cubook/model/task.dart';
 import 'package:cubook/model/themeInfo.dart';
-import 'package:cubook/task_detail_scout_confirm/taskDetailScoutConfirm_view.dart';
-import 'package:cubook/task_list_scout_confirm/taskListScoutConfirm_view.dart';
+import 'package:cubook/checkScoutTaskDetail/taskDetailScoutConfirm_view.dart';
+import 'package:cubook/checkScoutTaskList/taskListScoutConfirm_view.dart';
 import 'package:cubook/userDetail/userDetail_model.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class TabInfo {
-
   TabInfo(this.label, this.widget);
   String label;
   Widget widget;
 }
 
 class SelectBook extends StatelessWidget {
-
   SelectBook(String? uid) {
     this.uid = uid;
   }
@@ -30,7 +28,8 @@ class SelectBook extends StatelessWidget {
     return SingleChildScrollView(
         child: Column(
       children: <Widget>[
-        Consumer<UserDetailModel>(builder: (BuildContext context, UserDetailModel model, Widget? child) {
+        Consumer<UserDetailModel>(builder:
+            (BuildContext context, UserDetailModel model, Widget? child) {
           if (model.userSnapshot == null) {
             model.getSnapshot(uid);
           } else if (model.userSnapshot!.get('uid') != uid) {
@@ -57,9 +56,9 @@ class SelectBook extends StatelessWidget {
                               ),
                               onTap: () {
                                 if (task.getAllMap(type[index])!.length != 1) {
-                                  Navigator.push(context,
-                                      MaterialPageRoute<TaskListScoutConfirmView>(
-                                          builder: (BuildContext context) {
+                                  Navigator.push(context, MaterialPageRoute<
+                                          TaskListScoutConfirmView>(
+                                      builder: (BuildContext context) {
                                     return TaskListScoutConfirmView(
                                         type[index], uid);
                                   }));
@@ -81,34 +80,31 @@ class SelectBook extends StatelessWidget {
                                           right: 10),
                                       child: Selector<UserDetailModel,
                                               DocumentSnapshot?>(
-                                          selector: (BuildContext context, UserDetailModel model) =>
+                                          selector: (BuildContext context,
+                                                  UserDetailModel model) =>
                                               model.userSnapshot,
-                                          builder: (BuildContext context, DocumentSnapshot<Object?>? snapshot, Widget? child) => snapshot !=
-                                                  null
-                                              ? model.userData![type[index]] !=
-                                                      null
-                                                  ? CircularProgressIndicator(
-                                                      backgroundColor: theme.getIndicatorColor(
-                                                          type[index]),
-                                                      valueColor: const AlwaysStoppedAnimation<Color>(
-                                                          Colors.white),
-                                                      value: model.userData![type[index]].length /
-                                                          task
-                                                              .getAllMap(
-                                                                  type[index])!
-                                                              .length)
+                                          builder: (BuildContext context,
+                                                  DocumentSnapshot<Object?>?
+                                                      snapshot,
+                                                  Widget? child) =>
+                                              snapshot != null
+                                                  ? model.userData![type[index]] !=
+                                                          null
+                                                      ? CircularProgressIndicator(
+                                                          backgroundColor: theme
+                                                              .getIndicatorColor(type[index]),
+                                                          valueColor: const AlwaysStoppedAnimation<Color>(Colors.white),
+                                                          value: model.userData![type[index]].length / task.getAllMap(type[index])!.length)
+                                                      : CircularProgressIndicator(backgroundColor: theme.getIndicatorColor(type[index]), valueColor: const AlwaysStoppedAnimation<Color>(Colors.white), value: 0)
                                                   : CircularProgressIndicator(
-                                                      backgroundColor: theme.getIndicatorColor(type[index]),
-                                                      valueColor: const AlwaysStoppedAnimation<Color>(Colors.white),
-                                                      value: 0)
-                                              : CircularProgressIndicator(
-                                                  backgroundColor:
-                                                      theme.getIndicatorColor(
-                                                          type[index]),
-                                                  valueColor:
-                                                      const AlwaysStoppedAnimation<
-                                                          Color>(Colors.white),
-                                                ))),
+                                                      backgroundColor: theme
+                                                          .getIndicatorColor(
+                                                              type[index]),
+                                                      valueColor:
+                                                          const AlwaysStoppedAnimation<
+                                                                  Color>(
+                                                              Colors.white),
+                                                    ))),
                                   Padding(
                                       padding: const EdgeInsets.only(left: 10),
                                       child: Column(
@@ -122,7 +118,8 @@ class SelectBook extends StatelessWidget {
                                                   alignment:
                                                       Alignment.centerLeft,
                                                   child: Text(
-                                                    theme.getTitle(type[index])!,
+                                                    theme
+                                                        .getTitle(type[index])!,
                                                     style: const TextStyle(
                                                         fontSize: 23,
                                                         color: Colors.white),
